AI_ResNet/api_requirements.txt
Flask>=3.0.3
flask-cors>=4.0.1
torch==2.4.1
torchvision==0.19.1
pillow>=10.4.0
numpy>=1.26.4
